reorg test importing json
