--- test/asset-import/asset-import-json-build-commonjs/asset-import-json-build-commonjs.js
@@ -1,3 +1,0 @@
-import data from "./data.json"
-
-export const value = data
