--- test/asset-import/asset-import-json-build-commonjs/data.json
@@ -1,1 +1,0 @@
-42
