--- test/asset-import/asset-import-json-build-systemjs/asset-import-json-build-systemjs.test.js
@@ -1,48 +1,0 @@
-import { assert } from "@jsenv/assert"
-import {
-  resolveDirectoryUrl,
-  urlToRelativeUrl,
-  urlToBasename,
-} from "@jsenv/filesystem"
-
-import { buildProject } from "@jsenv/core"
-import { jsenvCoreDirectoryUrl } from "@jsenv/core/src/internal/jsenvCoreDirectoryUrl.js"
-import {
-  GENERATE_SYSTEMJS_BUILD_TEST_PARAMS,
-  IMPORT_SYSTEM_JS_BUILD_TEST_PARAMS,
-} from "@jsenv/core/test/TEST_PARAMS_BUILD_SYSTEMJS.js"
-import { browserImportSystemJsBuild } from "@jsenv/core/test/browserImportSystemJsBuild.js"
-
-const testDirectoryUrl = resolveDirectoryUrl("./", import.meta.url)
-const testDirectoryRelativeUrl = urlToRelativeUrl(
-  testDirectoryUrl,
-  jsenvCoreDirectoryUrl,
-)
-const testDirectoryname = urlToBasename(testDirectoryRelativeUrl)
-const jsenvDirectoryRelativeUrl = `${testDirectoryRelativeUrl}.jsenv/`
-const buildDirectoryRelativeUrl = `${testDirectoryRelativeUrl}dist/systemjs/`
-const mainFilename = `${testDirectoryname}.js`
-const entryPointMap = {
-  [`./${testDirectoryRelativeUrl}${mainFilename}`]: "./main.js",
-}
-await buildProject({
-  ...GENERATE_SYSTEMJS_BUILD_TEST_PARAMS,
-  jsenvDirectoryRelativeUrl,
-  buildDirectoryRelativeUrl,
-  entryPointMap,
-  minify: true,
-})
-const { namespace } = await browserImportSystemJsBuild({
-  ...IMPORT_SYSTEM_JS_BUILD_TEST_PARAMS,
-  testDirectoryRelativeUrl,
-  htmlFileRelativeUrl: "./index.html",
-})
-
-const actual = namespace
-const expected = {
-  default: {
-    whatever: "It's cool",
-    [`w"ow`]: 42,
-  },
-}
-assert({ actual, expected })
