--- test/asset-import/asset-import-json-launch-node/asset-import-json-launch-node.js
@@ -1,3 +1,0 @@
-import data from "./data.json"
-
-export default data
