--- test/asset-import/asset-import-json-launch-node/data.json
@@ -1,3 +1,0 @@
-{
-  "foo": true
-}
