+++ test/importing_json/main.html
@@ -2,8 +2,9 @@
   <head>
     <title>Test</title>
     <meta charset="utf-8" />
-    <script src="system.js"></script>
   </head>
 
-  <body></body>
+  <body>
+    <script type="module" src="./main.js"></script>
+  </body>
 </html>
