+++ test/importing_json/main.js
